x/airblogs/types/message_postblog_test.go
package types

import (
	"testing"

	"airblogs/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgPostblog_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgPostblog
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgPostblog{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgPostblog{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
